rename test/coverage package to test/coverage_scalatest
