+++ test/coverage_scalatest/B2.java
@@ -1,5 +1,5 @@
-package coverage;
-
+package coverage_scalatest;
+import coverage_scalatest.C2;
 class B2 {
     public static String b2_a() {
         return C2.c2("hello from b2_a");
