+++ test/coverage_scalatest/BUILD
