+++ test/coverage_scalatest/C2.scala
@@ -1,4 +1,4 @@
-package coverage;
+package coverage_scalatest
 
 object C2 {
   def c2(input: String): String =
