+++ test/coverage_scalatest/TestB2.java
@@ -1,7 +1,6 @@
-package coverage;
+package coverage_scalatest;
 
 import org.junit.Test;
-import org.junit.Assert.*;
 
 public class TestB2 {
