+++ test/coverage_scalatest/expected-coverage.dat
@@ -1,0 +1,106 @@
+SF:test/coverage_scalatest/A1.scala
+FN:-1,coverage_scalatest/A1$::<clinit> ()V
+FN:3,coverage_scalatest/A1$::<init> ()V
+FN:5,coverage_scalatest/A1$::a1 (Z)Lcoverage_scalatest/B1$;
+FN:-1,coverage_scalatest/A1::a1 (Z)Lcoverage_scalatest/B1$;
+FNDA:1,coverage_scalatest/A1$::<clinit> ()V
+FNDA:1,coverage_scalatest/A1$::<init> ()V
+FNDA:1,coverage_scalatest/A1$::a1 (Z)Lcoverage_scalatest/B1$;
+FNDA:0,coverage_scalatest/A1::a1 (Z)Lcoverage_scalatest/B1$;
+FNF:4
+FNH:3
+BA:5,2
+BRF:1
+BRH:1
+DA:3,1
+DA:5,4
+DA:6,1
+DA:7,4
+LH:4
+LF:4
+end_of_record
+SF:test/coverage_scalatest/A2.scala
+FN:-1,coverage_scalatest/A2$::<clinit> ()V
+FN:3,coverage_scalatest/A2$::<init> ()V
+FN:5,coverage_scalatest/A2$::a2 ()V
+FN:-1,coverage_scalatest/A2::a2 ()V
+FNDA:1,coverage_scalatest/A2$::<clinit> ()V
+FNDA:1,coverage_scalatest/A2$::<init> ()V
+FNDA:1,coverage_scalatest/A2$::a2 ()V
+FNDA:0,coverage_scalatest/A2::a2 ()V
+FNF:4
+FNH:3
+DA:3,1
+DA:5,12
+DA:8,4
+LH:3
+LF:3
+end_of_record
+SF:test/coverage_scalatest/B1.scala
+FN:-1,coverage_scalatest/B1$::<clinit> ()V
+FN:3,coverage_scalatest/B1$::<init> ()V
+FN:6,coverage_scalatest/B1$::not_called ()V
+FN:-1,coverage_scalatest/B1::not_called ()V
+FNDA:1,coverage_scalatest/B1$::<clinit> ()V
+FNDA:1,coverage_scalatest/B1$::<init> ()V
+FNDA:0,coverage_scalatest/B1$::not_called ()V
+FNDA:0,coverage_scalatest/B1::not_called ()V
+FNF:4
+FNH:2
+DA:3,1
+DA:6,0
+DA:9,4
+LH:2
+LF:3
+end_of_record
+SF:test/coverage_scalatest/B2.java
+FN:3,coverage_scalatest/B2::<init> ()V
+FN:5,coverage_scalatest/B2::b2_a ()Ljava/lang/String;
+FN:9,coverage_scalatest/B2::b2_b ()V
+FNDA:0,coverage_scalatest/B2::<init> ()V
+FNDA:1,coverage_scalatest/B2::b2_a ()Ljava/lang/String;
+FNDA:0,coverage_scalatest/B2::b2_b ()V
+FNF:3
+FNH:1
+DA:3,0
+DA:5,3
+DA:9,0
+DA:10,0
+LH:1
+LF:4
+end_of_record
+SF:test/coverage_scalatest/C2.scala
+FN:-1,coverage_scalatest/C2$::<clinit> ()V
+FN:3,coverage_scalatest/C2$::<init> ()V
+FN:5,coverage_scalatest/C2$::c2 (Ljava/lang/String;)Ljava/lang/String;
+FN:-1,coverage_scalatest/C2::c2 (Ljava/lang/String;)Ljava/lang/String;
+FNDA:1,coverage_scalatest/C2$::<clinit> ()V
+FNDA:1,coverage_scalatest/C2$::<init> ()V
+FNDA:1,coverage_scalatest/C2$::c2 (Ljava/lang/String;)Ljava/lang/String;
+FNDA:1,coverage_scalatest/C2::c2 (Ljava/lang/String;)Ljava/lang/String;
+FNF:4
+FNH:4
+DA:3,1
+DA:5,9
+DA:7,4
+LH:3
+LF:3
+end_of_record
+SF:test/coverage_scalatest/D1.scala
+FN:-1,coverage_scalatest/D1$::<clinit> ()V
+FN:3,coverage_scalatest/D1$::<init> ()V
+FN:5,coverage_scalatest/D1$::veryLongFunctionNameIsHereAaaaaaaaa ()V
+FN:-1,coverage_scalatest/D1::veryLongFunctionNameIsHereAaaaaaaaa ()V
+FNDA:1,coverage_scalatest/D1$::<clinit> ()V
+FNDA:1,coverage_scalatest/D1$::<init> ()V
+FNDA:1,coverage_scalatest/D1$::veryLongFunctionNameIsHereAaaaaaaaa ()V
+FNDA:0,coverage_scalatest/D1::veryLongFunctionNameIsHereAaaaaaaaa ()V
+FNF:4
+FNH:3
+DA:3,1
+DA:5,6
+DA:9,8
+DA:41,4
+LH:4
+LF:4
+end_of_record
