+++ test/shell/test_coverage_scalatest.sh
@@ -5,15 +5,15 @@ dir=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 runner=$(get_test_runner "${1:-local}")
 
 test_coverage_on() {
-    bazel coverage //test/coverage:test-scalatest
-    diff test/coverage/expected-coverage.dat $(bazel info bazel-testlogs)/test/coverage/test-scalatest/coverage.dat
+    bazel coverage //test/coverage_scalatest:test-scalatest
+    diff test/coverage_scalatest/expected-coverage.dat $(bazel info bazel-testlogs)/test/coverage_scalatest/test-scalatest/coverage.dat
 }
 
 test_coverage_includes_test_targets() {
     bazel coverage \
           --instrument_test_targets=True \
-          //test/coverage:test-scalatest
-    grep -q "SF:test/coverage/TestWithScalaTest.scala" $(bazel info bazel-testlogs)/test/coverage/test-scalatest/coverage.dat
+          //test/coverage_scalatest:test-scalatest
+    grep -q "SF:test/coverage_scalatest/TestWithScalaTest.scala" $(bazel info bazel-testlogs)/test/coverage_scalatest/test-scalatest/coverage.dat
 }
 
 $runner test_coverage_on
